Use separate file descriptor per session and close it immediately after reading
I misunderstood how gopasslib uses the file descriptor. I assumed
that it would need the file descriptor even after the initial
decoding. After reading the code, I understood that it reads and
decodes the entire file. This means the file descriptor doesn't
need to be kept around and we can close it immediately after
reading the database file.