Put back hack that reloads page after logout
The state reset should log out you since we destroyed the cookie,
but for some reason, after a second or something, it logs back in
and the app ends up in a broken state.

I don't have the energy to figure out why, so I am putting back
the hack to make it work.